+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Added `Get-vRLIContentPack` cmdlet to support install and upgrade of content packs from the vRealize Log Insight in-product marketplace.
 - Added `Get-vROpsLogForwardingConfig` cmdlet to support verification of vRealize Operations log forwarding.
 - Added `Get-vROpsLogForwarding` cmdlet to support verification of vRealize Operations log forwarding.
+- Added `Add-vROPSAdapterVr` cmdlet to support adding a vSphere Replication adapter to vRealize Operations.
 - Updated `Export-vRLIJsonSpec` cmdlet to define vRealize Log Insight v8.8.2 as the default install for VCF 4.5.
 - Updated `Export-vRAJsonSpec` cmdlet to define vRealize Automation v8.8.2 as the default install for VCF 4.5.
 - Updated Sample Scripts in the SampleScripts\ila folder, to support the VCF 4.5.x Planning and Prep Workbook: